849- string length generation fixes (without soft constraints)

+++ generator/src/main/java/com/scottlogic/deg/generator/generation/GenerationConfig.java
@@ -176,6 +176,6 @@ public class GenerationConfig {
 
         public static final BigDecimal NUMERIC_MAX = new BigDecimal("1e20");
         public static final BigDecimal NUMERIC_MIN = new BigDecimal("-1e20");
-        public static final BigDecimal MAX_STRING_LENGTH = BigDecimal.valueOf(1000);
+        public static final int MAX_STRING_LENGTH = 1000;
     }
 }
